requirements.txt
numpy
onnx
opencv-python
pandas
pillows
scikit-learn
scipy
torch
tqdm
openai